fix: Force white for icon animation to password view
Password view is not compatible with dark mode but it was
using for the icon animation a color which is since recent commits
responsive to dark mode.

So here we force the icon animation color to white to stay consistent
and one day we may convert totally the password view to dark mode.