fix: Force white for icon animation to password view
Password view is loaded from the cloudery. We want to display the
icon animation in white if it is a Cozy cloudery and in primary color
if it is a GL cloudery.

We used previously `paperBackgroundColor` for Cozy cloudery which was
always white. But now `paperBackgroundColor` react to color scheme and
we do not want it. So let's force white.