feat: Enable CacheMode for CozyProxyWebView on android when offline
When a cozy-app's WebView is served offline, then the cozy-app bundle
is served from local folder using the CozyProxyWebview, and since
previous commits, the cozy-client's queries are served through
`FlagshipLink`

With those mechanisms, nearly all data are available offline except two
things:
- cozy-stack static assets (i.e. some css files, avatar, partners logos
etc)
- `fetchJSON()` requests

This commit tries to handle the static assets part

Those assets are often served with HTTP cache activated, and so the
WebView should be able to load them correctly when the device is
offline. However on Android, the WebView's cache seems not to persist
after the application is closed

Activating the `cacheMode=LOAD_CACHE_ELSE_NETWORK` seems to fix this
behavior. But it prioritizes too much the cache over fresh data, so we
want to activate it only when strictly necessary, which is when the
application is offline

This also allows to load partners logos that are not cached by the
cozy-stack (as of today). But we expect to modify the cozy-stack to
enable cache on them