fix: Update build patch for react-native-background-upload
react-native-background-upload is not maintained. We already had a
patch for react-native 0.72, but we are getting an error when
uploading a file when targeting Android 34 : "One of
RECEIVER_EXPORTED or RECEIVER_NOT_EXPORTED should be specified when a
receiver isn't being registered exclusively for system broadcasts".

Fix is done in android-upload-service
https://github.com/gotev/android-upload-service/issues/642

New patch is taken from
https://github.com/Vydia/react-native-background-upload/pull/347