fix: Set CryptoWebView as ready only when it is really ready
We use react-native-webview onLoadEnd callback to set the
CryptoWebView as ready. It is an important action because other
webviews are rendered only when the CryptoWebView is ready. But the
onLoadEnd can be triggered before the javascript, where we add
our crypto functions, is executed.
It is also executed two times (I do not understand why). So here I
send a post message from the CryptoWebView to indicate that it is
ready.

It fixes a race condition we may encounter more and more.