feat: Allow the user to choose to save debug traces on error
This feature is still behind the `clisk.automatic.html-on-error` flag
because it still needs some UI work.

When an unexpected error is raised from a clisk konnector, the current
page in the worker is displayed to user and a button to send traces is
proposed which does exactly the same as the `clisk.html-on-error` flag.