feat: Speed up app display
We used to init search at app initilization. However, we noticed that
this can be time consuming and slow down the app openening. Thus, we
prefer to make the user wait for an extra few seconds for the first
search, to init the search at this moment, rather than waiting at each
app opening.  Note we previously have added the index persistence
feature, in order to speed up this initialization.