feat: Init search at app startup, with custom delay
The search index are built after a replication is complete. So we want
to start the replication soon enough to have the search available for
the user. But it cannot be started immediatly as it can be a heavy
process, especially for first replication, and it can slow down the app
startup.
So we delay the replication with what appears as an acceptable delay,
i.e. 10s. From our experiments, we noticed a 7s startup time on a slow
device with an instance including 25K+ files.