fix: Redirection is not properly working in offline
To compute the home redirection link at startup, we query the instance
settings doc to extract the `default_redirection` attribute.
However, we were using the `data.attributes` object to access it, which
cannot work in offline, because there is no JSON-API formatting coming
from PouchDB. And it is not required with online, because cozy-client
does the JSON-API normalization.