feat: Upload log files to the Cozy instead of only sending them by email
The email intent seems not to be consistent, sometimes the intent does
not start on Android, sometimes it opens and then it closes directly on
iOS

For now we don't have any solution yet for this, but sending the email
is not the critical part for those tools. Uploading the files into the
user's Cozy would be enough

We already do that for Performances and Databases email reporters, but
we did not implement this yet on the Logs email reporter yet

This was done like this because we want to support sending logs even if
the user is not connected yet (and so we may not have any Cozy instance
identified to upload the files to)

But we can still improve this by uploading the log files to the Cozy
when the user is connected. When not connected we would still use the
email only solution

Note that for the upload of the Logs files, we first create a copy in a
temporary folder. This is because the logger can still add logs into
the log files while we try to send them. When this happens, we can get
some Stream errors due to the file size that changes while being read
for the upload. By copying the file first, we can easily prevent this
error