feat: Change transition between apps
Twake UI is moving from "icon growing to fullscreen page" animation as code to "fullscreen page with icon doing things" animation as lottie file. There is a default animation with Twake logo for new apps and settings.

Because now it is a fullscreen page, there is no more need to managing the first half or the possible premature end of the transition if the webview loads fast.