feat: Implement new Cloudery configuration for twake
We want to adapt the onboarding screens in order to use the new
Twake SSO login and subscription mechanisms

To allow this we introduce new "Cloudery" environments in order to
declare SSO urls

For now we want to keep the old Cozy onboarding screens for testing
purpose, so instead of replacing the environment strings we chose to
duplicate them

We also added a new overridable configuration to allow switching
between the new Twake and the old Cozy onboarding processes. The
default environment is the Twake onboarding

Note the `DeeplinksParserService` only refers to the old Cozy
environment strings, this is because those are used to handle the
login/onboard by email scenario that uses universal links. This
scenario does not exist anymore with Twake SSO processes, so there was
no need to adapt the condition