feat: Implement new Welcome screen for Twake SSO
We want to adapt the onboarding screens in order to use the new
Twake SSO login and subscription mechanisms

To allow this we rework the Welcome screen so it uses the new Twake UI
and so its login and signup buttons triggers the SSO process

Contrary to previous implementation that used WebViews, the new SSO
process is fully based on the InAppBrowser

The side effect is that we now prefer to use Schemes instead of
UniversalLinks. This is because UniversalLinks would not trigger
correctly in some scenario like when creating a Twake instance. In that
scenario, a few seconds can occure between the last user click and the
final redirection. In some devices (all iOS but also some Android
devices), the UniversalLink redirection would be prevented by the
browser's rules as it wouldn't be considered as a user action. So the
Schemes are prefered here as they would always be triggered. The only
cons is that Schemes would requires user's approval on Android, so the
user would have to validate a popup just before the final redirection

The Welcome screen also contains a "Use your company server" that would
allow old Cozy's not configured with Twake SSO to be still accessible

However we still did not implement the other OIDC scenario (on premise
instance, or partners). This will be implemented in the following weeks