feat: Implement OIDC login for new Twake scenario
When clicking on `Use your company server` button, we wan't to be able
to either:
- login to old Cozy instances (already implemented)
- login to OIDC instance (to be implemented)

This commit implement the detection of the login mechanism (classic vs
OIDC)

Then for OIDC instances, in implements the login mechanism. To make
this login work, we use the `client.authorize()` mechanism that already
implement in-app-browser based authorisation scenario that is able to
redirect to the OIDC portal if detected. This redirection is handled by
the cozy-stack