feat: Update op-sqlite to 14.1.4 to support 16KB
In 2764992, I added op-sqlite as a dependency to the same version
that the version used by pouchdb-adapter-react-native-sqlite. But
this version is not compatible with 16KB page size so here I update
it to 14.1.4 which is compatible and force this version to
pouchdb-adapter-react-native-sqlite.

I did not see breaking changes affecting
pouchdb-adapter-react-native-sqlite.